[refactor/#12] 페이지 폴더 내로 파일 이동
