+++ src/pages/Login/Input.css
@@ -22,7 +22,7 @@
 .CustomInput {
   width: 560px;
   height: 48px;
-  padding: 16px 404px 15px 15px;
+  padding: 16px 0px 15px 15px;
   align-items: center;
   flex-shrink: 0;
   border-radius: 8px;
