--- src/pages/Signup/Login.tsx
@@ -1,5 +1,0 @@
-const Login = () => {
-  return <div>로그인</div>;
-};
-
-export default Login;
