--- src/pages/Signup/Signup.tsx
@@ -1,5 +1,0 @@
-const Signup = () => {
-  return <div>회원가입</div>;
-};
-
-export default Signup;
