+++ src/routes/Router.tsx
@@ -1,21 +1,21 @@
 import { createBrowserRouter } from "react-router-dom";
 import Home from "@/pages/Home/Home";
 import MyPage from "@/pages/MyPage/MyPage";
-import Oauth from "@/pages/Signup/Oauth";
-import Signup from "@/pages/Signup/Signup";
+import Oauth from "@/pages/Login/Oauth";
+import SignPageOne from "@/pages/Login/SignPageOne";
 import ProtectedRoute from "@/components/Common/ProtectedRouter";
 import MainLayout from "@/layouts/MainLayout";
-import Login from "@/pages/Signup/Login";
+import LoginPage from "@/pages/Login/LoginPage";
 
 export const router = createBrowserRouter([
   {
     path: "/",
-    element: <Login />,
+    element: <LoginPage />,
   },
   { path: "oauth", element: <Oauth /> },
   {
     path: "signup",
-    element: <Signup />,
+    element: <SignPageOne />,
   },
   {
     path: "user",
